添加从数据文件读取帽子偏移

+++ CreateItems/createItems.py
@@ -1,14 +1,22 @@
+import os
 import json
 import requests
 
-#url = "https://api.steampowered.com/IGameInventory/GetItemDefArchive/v0001?appid=3419430&digest=CEAA2C94E932B799E7B7498D18EB450004622927"
+destination_path = "..\\Assets\\"
+
+# url = "https://api.steampowered.com/IGameInventory/GetItemDefArchive/v0001?appid=3419430&digest=CEAA2C94E932B799E7B7498D18EB450004622927"
 url = "https://api.steampowered.com/IGameInventory/GetItemDefArchive/v0001?appid=3419430&digest=D93DA1D3EDE56235556DA16C5C2A5DC9303B0EC8"
 
-response = requests.get(url, verify=False)
-content = response.text.rstrip(b'\x00'.decode())
-if response.status_code == 200:
-    with open("inventory.json", "w") as file:
-        file.write(content)
+content = ""
+if os.path.exists("inventory.json"):
+    with open("inventory.json", "r") as file:
+        content = file.read()
+else:
+    response = requests.get(url, verify=False)
+    content = response.text.rstrip(b'\x00'.decode())
+    if response.status_code == 200:
+        with open("inventory.json", "w") as file:
+            file.write(content)
 
 json_data = json.loads(content)
 
@@ -19,12 +27,10 @@ for item in json_data:
         data = {
             "name": item.get("name"),
             "description": item.get("description"),
-            "Timestamp": item.get("Timestamp"),
             "modified": item.get("modified"),
-            "date_created": item.get("date_created"),
+            "created": item.get("date_created"),
             "icon": item.get("icon_url").split("/")[-1],
             "image": "",
-            "name_color": item.get("name_color"),
             "tags": item.get("tags").replace(";itemslot:skin", "").replace(";itemslot:hat", ""),
         }
         if item.get("item_slot") == "skin":
@@ -45,5 +51,5 @@ newdata = {
     "hat": hatdata
 }
 
-with open("items.json", "w") as json_file:
-    json.dump(newdata, json_file, indent=2)
+with open(destination_path + "items.json", "w") as json_file:
+    json.dump(newdata, json_file)
